requirements.txt
numpy==1.25.2
pandas==2.1.0
zarr==2.16.1
geopy==2.4.0
xarray==2023.8.0
dask==2023.6.0
pyproj==3.5.0
geojson==2.5.0
setuptools==60.0.0
cachetools==4.1.1
ipython==8.11.0
requests==2.31.0
aiohttp==3.8.5
nest_asyncio==1.4.1
typing-extensions==4.6.1
wheel==0.36.
# plotly-dash-vis
matplotlib==3.8.2
plotly==5.18.0
dash==2.16.1
dash-auth==2.0.0
dash-extensions==1.0.4
dash-bootstrap-components==1.5.0
dash-iconify==0.1.2
gunicorn==20.1.0
flask-login==0.6.2
CairoSVG==2.5.1
Werkzeug==2.3.8
Hypercorn==0.16.0
# testing
pytest-cov==4.0.0
pytest==7.2.2
mock==5.0.1
fastapi==0.103.1
pydantic==2.3.0
uvicorn==0.23.2